+++ lib/jsonApi.js
@@ -5,7 +5,7 @@ var _ = require("underscore");
 var Joi = require("joi");
 var router = require("./router.js");
 var responseHelper = require("./responseHelper.js");
-var handlers = require("./handlers.js");
+var routes = require("./routes");
 var mockHandlers = require("./mockHandlers.js");
 var postProcess = require("./postProcess.js");
 
@@ -20,7 +20,6 @@ jsonApi.setConfig = function(apiConfig) {
   responseHelper.setMetadata(jsonApi._apiConfig.meta);
   router.using(jsonApi);
   postProcess.using(jsonApi);
-  handlers.using(jsonApi);
 };
 
 jsonApi._cleanBaseUrl = function(base) {
@@ -81,17 +80,7 @@ jsonApi.define = function(resourceConfig) {
 };
 
 jsonApi.start = function() {
-  handlers._registerForeignKeySearch();
-  handlers._registerSearch();
-  handlers._registerFind();
-  handlers._registerRelated();
-  handlers._registerRelationships();
-  handlers._registerCreate();
-  handlers._registerDelete();
-  handlers._registerUpdate();
-  handlers._registerUpdateRelation();
-  handlers._registerAddRelation();
-  handlers._registerRemoveRelation();
+  routes.register();
   router.listen(jsonApi._apiConfig.port);
 };
