+++ lib/router.js
@@ -36,7 +36,7 @@ router.using = function(jsonApi) {
 };
 
 router.bindToServer = function(config, callback) {
-  app[config.verb](config.path, function(req, res) {
+  app[config.verb](router._jsonApi._apiConfig.base + config.path, function(req, res) {
     var request = router._getParams(req);
     var resourceConfig = router._jsonApi._resources[request.params.type];
     router._setResponseHeaders(request, res);
